Delete statement about expert review
This statement is technically factual but highly misleading in this
context as far as implying endorsement of this crate. As far as I can
tell from the linked thread, no seasoned Rust expert has said "yes this
crate is sound". Me as one of those experts basically lost interest
after discovering a large number of ways to trigger UB via the safe API.