Add third from_fn constructor type
This allows much more flexible construction. But comes at the cost of not
allowing the automatic Clone impl. It ought to be safe on current Rust, but we
might have to revise this specific constructor with future Rust versions, when
expressing a closure that captures a reference that outlives the hrtb function
signature ref it's passed into becomes possible.