Consolidate constructor types
This reduces the available constructor types from 4 to 2, only leaving the
closure based constructors. These are more flexible and ergonomic to use in most
cases. They are renamed to new and try_new making the interface more consistent.
Plus this significantly cuts down on implementation complexity as the
constructor logic has to be repeated. Further this removes the need of having
users commit to a constructor type upfront. They can choose the one they need,
even different ones in different circumstances.

In addition the documentation and examples have been reworked.