Fix "best-effort" typo.
"Effort" was missing the letter "t". I've also slightly tweaked the wording in a way that reads better to me, but obviously feel free to change it however you see fit :)

I am a bit confused about the statement in general, though: is MSRV truly best-effort, or is it guaranteed to only change in major versions? Perhaps more clarification would be good.

In any case, this crate is great and I look forward to using it!